fight

+++ app/src/main/java/ru/kompod/moonlike/data/network/mapper/MapMapper.kt
@@ -6,7 +6,6 @@ package ru.kompod.moonlike.data.network.mapper
 import android.content.Context
 import ru.kompod.moonlike.data.network.model.MapApiModel
 import ru.kompod.moonlike.data.network.model.TravelApiModel
-import ru.kompod.moonlike.data.repository.AssetRepository
 import ru.kompod.moonlike.domain.entity.base.MapObject
 import ru.kompod.moonlike.domain.entity.base.MonsterObject
 import ru.kompod.moonlike.domain.entity.base.TravelObject
@@ -20,17 +19,17 @@ class MapMapper @Inject constructor(
 ) {
     companion object {
         private const val BIOME_FIELD = "field"
-        private const val BIOME_FIELD_ID: Short = 1
+        private const val BIOME_FIELD_ID: Int = 1
         private const val BIOME_SWAMP = "swamp"
-        private const val BIOME_SWAMP_ID: Short = 2
+        private const val BIOME_SWAMP_ID: Int = 2
         private const val BIOME_MINE = "mine"
-        private const val BIOME_MINE_ID: Short = 3
+        private const val BIOME_MINE_ID: Int = 3
         private const val BIOME_RUIN = "ruin"
-        private const val BIOME_RUIN_ID: Short = 4
+        private const val BIOME_RUIN_ID: Int = 4
         private const val BIOME_DESERT = "desert"
-        private const val BIOME_DESERT_ID: Short = 5
+        private const val BIOME_DESERT_ID: Int = 5
         private const val BIOME_SNOW = "snow"
-        private const val BIOME_SNOW_ID: Short = 6
+        private const val BIOME_SNOW_ID: Int = 6
     }
 
     fun mapMapApiModelToEntity(model: MapApiModel, monsters: List<MonsterObject>): MapObject = MapObject(
@@ -47,7 +46,7 @@ class MapMapper @Inject constructor(
         actors = model.actors
     )
 
-    private fun mapBiomeByTitle(biome: String): Short =
+    private fun mapBiomeByTitle(biome: String): Int =
         when (biome) {
             BIOME_FIELD -> BIOME_FIELD_ID
             BIOME_SWAMP -> BIOME_SWAMP_ID
