+++ app/src/main/java/ru/kompod/moonlike/data/network/model/RoleApiModel.kt
@@ -8,17 +8,17 @@ import ru.kompod.moonlike.utils.NO_ID
 import ru.kompod.moonlike.utils.extensions.kotlin.empty
 
 data class RoleApiModel(
-    val id: Short = NO_ID,
+    val id: Int = NO_ID,
     val label: String = String.empty,
     val description: String = String.empty,
     val states: States = States()
 )
 
 data class States(
-    val hp: Short = 30,
-    val sp: Short = 0,
-    @SerializedName("f-atk") val fAtk: Short = 0,
-    @SerializedName("f-def") val fDef: Short = 0,
-    @SerializedName("m-atk") val mAtk: Short = 0,
-    @SerializedName("m-def") val mDef: Short = 0
+    val hp: Int = 30,
+    val sp: Int = 0,
+    @SerializedName("f-atk") val fAtk: Int = 0,
+    @SerializedName("f-def") val fDef: Int = 0,
+    @SerializedName("m-atk") val mAtk: Int = 0,
+    @SerializedName("m-def") val mDef: Int = 0
 )
